refactor :: processManager 분리
